Re-structure things to avoid import cycles.
